Reorganize benchmarks to enable future code reuse
