+++ src/test_utils/mod.rs
@@ -1,5 +1,4 @@
 use borsh::{BorshDeserialize, BorshSerialize};
-use criterion::criterion_main;
 use near_primitives_core::config::VMConfig;
 use near_primitives_core::contract::ContractCode;
 use near_primitives_core::profile::ProfileData;
@@ -13,20 +12,11 @@ use primitive_types::U256;
 use rlp::RlpStream;
 use secp256k1::{self, Message, PublicKey, SecretKey};
 
-use aurora_engine::parameters::{NewCallArgs, SubmitResult};
-use aurora_engine::prelude::Address;
-use aurora_engine::storage;
-use aurora_engine::transaction::{EthSignedTransaction, EthTransaction};
-use aurora_engine::types;
-
-#[cfg(feature = "profile_eth_gas")]
-use aurora_engine::prelude::ETH_GAS_USED;
-
-mod eth_deploy_code;
-mod eth_erc20;
-mod eth_standard_precompiles;
-mod eth_transfer;
-mod solidity;
+use crate::parameters::{NewCallArgs, SubmitResult};
+use crate::prelude::Address;
+use crate::storage;
+use crate::transaction::{EthSignedTransaction, EthTransaction};
+use crate::types;
 
 near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
     EVM_WASM_BYTES => "release.wasm"
@@ -34,12 +24,7 @@ near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
 
 pub const SUBMIT: &str = "submit";
 
-criterion_main!(
-    eth_deploy_code::benches,
-    eth_erc20::benches,
-    eth_standard_precompiles::benches,
-    eth_transfer::benches
-);
+pub mod solidity;
 
 pub struct AuroraRunner {
     pub aurora_account_id: String,
